FAIR-1779 删除不必要的import
